+++ src/components/TemplateNavbar.jsx
@@ -2,7 +2,7 @@ import Header from '@/components/Header'
 import Nav from '@/components/Nav'
 import styles from '@/styles/modules/Template.module.css'
 
-export default function TemplateHead({ children }) {
+export default function TemplateNavBar({ children }) {
 
     return (
         <main className={styles.main_navBar}>
