+++ src/pages/_app.js
@@ -1,10 +1,10 @@
-import TemplateHead from '@/components/TemplateNavbar'
+import TemplateNavbar from '@/components/TemplateNavbar'
 import '@/styles/globals.css'
 
 export default function App({ Component, pageProps }) {
-  return (
-    <TemplateHead>
-      <Component {...pageProps} />
-    </TemplateHead>
-  )
+    return (
+        <TemplateNavbar>
+            <Component {...pageProps} />
+        </TemplateNavbar>
+    )
 }
