+++ src/styles/modules/OrderList.module.css
@@ -1,46 +1,87 @@
 .container_order {
-  display: flex;
-  width: 50%;
-  justify-content: end;
+    display: flex;
+    width: 50%;
+    justify-content: end;
 }
 
 .container_order .list_order {
-  width: 65%;
-  display: flex;
-  flex-direction: column;
-  gap: 1em;
-  background-color: #fff;
-  color: #000;
-  border-radius: 8px;
-  padding: 2em;
+    width: 65%;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    gap: 1em;
+    background-color: #fff;
+    color: #000;
+    border-radius: 8px;
 }
 
-.container_order .list_order .title {
-  font-size: 1.5em;
-  font-weight: 600;
+.container_order .title {
+    font-size: 1.5em;
+    font-weight: 600;
+}
+
+.container_order .sublist_order {
+    display: flex;
+    gap: 1;
+    flex-direction: column;
+    box-shadow: 0px 3px 16px 0px rgba(25, 25, 25, 0.2);
+    padding: 2em;
+}
+
+.container_order .list_scroll {
+    display: flex;
+    flex-direction: column;
+    padding: 2em;
+    overflow-y: scroll;
+    height: 500px;
+}
+
+.container_order button {
+    margin-top: 1em;
+    border-radius: 8px;
+    width: 100%;
+    height: 1em;
+    padding: 1em 0;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: #fff;
+    font-size: 1.2em;
+    font-weight: 500;
+    background-color: #41d635;
+    cursor: pointer;
+    transition: 0.2s;
+}
+
+.container_order button:hover {
+    transform: scale(1.01);
+}
+
+.container_order button:active {
+    transform: scale(1);
 }
 
 .container_item {
-  display: flex;
-  flex-direction: column;
+    display: flex;
+    flex-direction: column;
 }
 
 .container_item .item {
-  display: flex;
-  justify-content: space-between;
+    display: flex;
+    justify-content: space-between;
 }
 
 .container_item .item .amount {
-  display: flex;
-  gap: 1em;
+    display: flex;
+    gap: 1em;
 }
 
 .container_item .item.gray {
-  color: #828282;
+    color: #828282;
 }
 
 .container_item .line {
-  width: 100%;
-  height: 2px;
-  background-color: #000;
+    width: 100%;
+    height: 2px;
+    background-color: #000;
 }
