+++ src/styles/modules/Template.module.css
@@ -1,34 +1,34 @@
 .main_navBar {
-  display: flex;
+    display: flex;
 }
 
 .main {
-  width: 100%;
-  margin-top: 8svh;
-  color: white;
-  padding: 2em 3em;
+    width: 100%;
+    margin-top: 8svh;
+    color: white;
+    padding: 2em 3em;
 }
 
 .main_body {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    height: 100%;
 }
 
 .name_page {
-  font-size: 2em;
-  font-weight: 600;
-  text-transform: uppercase;
-  margin-bottom: 0.5em;
+    font-size: 2em;
+    font-weight: 600;
+    text-transform: uppercase;
+    margin-bottom: 0.5em;
 }
 
 .container_content {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  gap: 1em;
-  background-color: #002650;
-  padding: 1.5em;
-  border-radius: 1em;
+    width: 100%;
+    height: 100%;
+    display: flex;
+    gap: 1em;
+    background-color: #002650;
+    padding: 1.5em;
+    border-radius: 1em;
 }
